Added a refresh button to sync latest efforts for the projects in project details page
